sched/requirements.txt
azureml-sdk==1.0.2
